test utils: move code for wait-for-pod-status to a new file
The client.go file was getting a little big. The helpers functions
that we use to check for pod status(es) get moved to a file dedicated
to checking on the status of things.
